+++ src/main/java/com/fibermc/essentialcommands/commands/joinpoints/JoinpointListCommand.java
@@ -1,4 +1,4 @@
-package com.fibermc.essentialcommands.commands;
+package com.fibermc.essentialcommands.commands.joinpoints;
 
 import java.util.*;
 import java.util.concurrent.CompletableFuture;
